Moved load(s) and dump(s) to an io package in preparation for plugins
