+++ src/pyfdl/__init__.py
@@ -1,7 +1,3 @@
-import json
-
-from typing import IO, Union
-
 from .common import (
     Base,
     Dimensions,
@@ -22,6 +18,7 @@ from .context import Context
 from .canvas_template import CanvasTemplate
 from .pyfdl import FDL
 from .errors import FDLError, FDLValidationError
+from .io.builtin import load, loads, dump, dumps
 
 __all__ = [
     'Base',
@@ -30,6 +27,8 @@ __all__ = [
     'Context',
     'DEFAULT_ROUNDING_STRATEGY',
     'Dimensions',
+    'dump',
+    'dumps',
     'FDL',
     'FDLError',
     'FDLValidationError',
@@ -48,69 +47,3 @@ __all__ = [
 ]
 
 __version__ = "0.1.0.dev0"
-
-
-def load(fp: IO, validate: bool = True) -> FDL:
-    """
-    Load an FDL from a file.
-
-    Args:
-        fp: file pointer
-        validate: validate incoming json with jsonschema
-
-    Raises:
-        jsonschema.exceptions.ValidationError: if the contents doesn't follow the spec
-
-    Returns:
-        FDL:
-    """
-    raw = fp.read()
-    return loads(raw, validate=validate)
-
-
-def loads(s: str, validate: bool = True) -> FDL:
-    """Load an FDL from string.
-
-    Args:
-        s: string representation of an FDL
-        validate: validate incoming json with jsonschema
-
-    Returns:
-        FDL:
-
-    """
-    fdl = FDL.from_dict(json.loads(s))
-
-    if validate:
-        fdl.validate()
-
-    return fdl
-
-
-def dump(obj: FDL, fp: IO, validate: bool = True, indent: Union[int, None] = 2):
-    """Dump an FDL to a file.
-
-    Args:
-        obj: object to serialize
-        fp: file pointer
-        validate: validate outgoing json with jsonschema
-        indent: amount of spaces
-    """
-    fp.write(dumps(obj, validate=validate, indent=indent))
-
-
-def dumps(obj: FDL, validate: bool = True, indent: Union[int, None] = 2) -> str:
-    """Dump an FDL to string
-
-    Args:
-        obj: object to serialize
-        validate: validate outgoing json with jsonschema
-        indent: amount of spaces
-
-    Returns:
-        string: representation of the resulting json
-    """
-    if validate:
-        obj.validate()
-
-    return json.dumps(obj.to_dict(), indent=indent, sort_keys=False)
